requirements.txt
openai==0.25.0
pandas==1.5.2
python-dotenv==0.21.0
streamlit==1.16.0
streamlit_extras==0.2.4
